Move trampoline code into trampoline package
