+++ bundle/trampoline/python_dbr_warning.go
@@ -1,4 +1,4 @@
-package python
+package trampoline
 
 import (
 	"context"
