+++ bundle/trampoline/trampoline_test.go
@@ -1,4 +1,4 @@
-package mutator
+package trampoline
 
 import (
 	"context"
